gui enhanced

+++ Modules/Bundles/org.mitk.gui.qt.diffusionimaging/src/internal/QmitkControlVisualizationPropertiesView.h
@@ -111,7 +111,7 @@ protected slots:
 
   void SetInteractor();
 
-  void PFWidth(double);
+  void PFWidth(int);
   void PFColor();
 
   void Heatmap();
